deployment/services/ols_profile/requirements.txt
# requirements.txt

fastapi==0.95.2
pydantic==1.10.7
motor==3.1.2
redis==4.5.5
typing==3.7.4.3
asyncio==3.4.3
uvicorn[standard]==0.22.0
strawberry-graphql[fastapi]==0.183.5
strawberry-graphql[cli]==0.183.5
